requirements.txt
flask_restful
flask-cors
pytest
pytest-mock
python-dotenv
pytest-dotenv
psycopg2-binary
pytest-env
flask-sqlalchemy
flask-restless
flask-oauthlib
